shore : 댓글 삭제 권한 에러 추가

+++ src/main/java/TubeSlice/tubeSlice/domain/comment/CommentService.java
@@ -36,8 +36,13 @@ public class CommentService {
     }
 
     @Transactional
-    public ApiResponse<SuccessStatus> deleteComment(Long commentId){
+    public ApiResponse<SuccessStatus> deleteComment(User user, Long commentId){
         Comment comment = commentRepository.findById(commentId).orElseThrow(()-> new CommentHandler(ErrorStatus.COMMENT_NOT_FOUND));
+
+        if(comment.getUser() != user){
+            throw new CommentHandler(ErrorStatus.COMMENT_FORBIDDEN);
+        }
+
         commentRepository.delete(comment);
 
         return ApiResponse.onSuccess(SuccessStatus._OK);
